authhelper: Deduplicate the handler used by BBA and CSA
